Subset of the AneuRisk65 benchmark data set

A data set containing the first derivative of the three-dimensional coordinates of the centerline of the internal carotid artery of 65 patients.

aneurisk65

Format

A list with 2 components:

x

A 65 x 1380 matrix containing, in each row, the evaluation grid for each patient;

y

A 65 x 3 x 1380 array containing, in each row, the values of the first derivative of each of the 3D coordinates of the ICA centerline, stored by a row in a matrix.

Source

This is a subset of the AneuRisk65 benchmark data set provided by the AneuRisk project.

References

Sangalli, L.M., Secchi, P. and Vantini, S. (2014), AneuRisk65: A dataset of three-dimensional cerebral vascular geometries, Electronic Journal of Statistics, 8 (2), 1879-1890.

Sangalli, L.M., Secchi, P. and Vantini, S. (2014), Analysis of AneuRisk65 data: K-mean Alignment, Electronic Journal of Statistics, 8 (2), 1891-1904.