Multiple CAPS mcaps objects

This function searches for clusters in the input data set using different strategies and generates an object of class mcaps which stores multiple objects of class caps. This is a helper function to facilitate comparison of clustering methods and choice of an optimal one.

compare_caps(
  x,
  y,
  n_clusters_max = 5L,
  metric = c("l2", "pearson"),
  clustering_method = c("kmeans", "hclust-complete", "hclust-average", "hclust-single"),
  warping_class = c("affine", "dilation", "none", "shift", "srsf"),
  centroid_type = c("mean", "medoid")
)

Arguments

x

A numeric matrix of shape \(N \times M\) specifying the grids of size \(M\) on which each of the \(N\) curves have been observed.

y

A numeric array of shape \(N \times L \times M\) specifying the \(N\)-sample of \(L\)-dimensional curves observed on grids of size \(M\).

n_clusters_max

An integer value specifying the maximum number of clusers to use. Defaults to 5L.

metric

A string specifying the metric used to compare curves. Choices are "l2" or "pearson". Defaults to "l2". This is used only when warping_class != "srsf".

clustering_method

A character vector specifying one or more clustering methods to be fit. Choices are "kmeans", "hclust-complete", "hclust-average" or "hclust-single". Defaults to all of them.

warping_class

A character vector specifying one or more classes of warping functions to use for curve alignment. Choices are "affine", "dilation", "none", "shift" or "srsf". Defaults to all of them.

centroid_type

A character vector specifying one or more ways to compute centroids. Choices are "mean" or "medoid". Defaults to all of them.

Value

An object of class mcaps which is a tibble::tibble storing the objects of class caps in correspondence of each combination of possible choices from the input arguments.

Examples

out <- compare_caps(
  x = simulated30$x,
  y = simulated30$y,
  n_clusters_max = 2,
  warping_class = "affine",
  clustering_method = "hclust-complete",
  centroid_type = "mean"
)