Distance Matrix for Functional Data with Amplitude and Phase Separation

This function computes the matrix of pairwise distances between curves a functional data sample. This can be achieved with or without phase and amplitude separation, which can be done using a variety of warping classes.

Usage

fdadist(
  x,
  y,
  warping_class = c("affine", "dilation", "none", "shift", "srsf"),
  metric = c("l2", "pearson"),
  cluster_on_phase = FALSE,
  labels = NULL
)

Arguments

x

A numeric matrix of shape \(N \times M\) specifying the grids of size \(M\) on which each of the \(N\) curves have been observed.

y

A numeric array of shape \(N \times L \times M\) specifying the \(N\)-sample of \(L\)-dimensional curves observed on grids of size \(M\).

warping_class

A string specifying the warping class Choices are "affine", "dilation", "none", "shift" or "srsf". Defaults to "affine". The SRSF class is the only class which is boundary-preserving.

metric

A string specifying the metric used to compare curves. Choices are "l2" or "pearson". Defaults to "l2". This is used only when warping_class != "srsf".

cluster_on_phase

A boolean specifying whether clustering should be based on phase variation or amplitude variation. Defaults to FALSE which implies amplitude variation.

labels

A character vector specifying curve labels. Defaults to NULL which uses sequential numbers as labels.

Value

A stats::dist object storing the distance matrix between the input curves using the metric specified through the argument metric and the warping class specified by the argument warping_class.

Examples

D <- fdadist(simulated30$x, simulated30$y)